--- frontend/src/components/IssueV1/components/Dialog/IssueReviewActionDialog/index.ts
@@ -1,3 +1,0 @@
-import IssueReviewActionDialog from "./ReviewDialog.vue";
-
-export default IssueReviewActionDialog;
