--- frontend/src/components/IssueV1/components/Dialog/IssueStatusActionDialog/index.ts
@@ -1,3 +1,0 @@
-import IssueStatusActionDialog from "./Dialog.vue";
-
-export default IssueStatusActionDialog;
