--- frontend/src/components/IssueV1/components/Dialog/TaskRolloutActionDialog/index.ts
@@ -1,3 +1,0 @@
-import TaskRolloutActionDialog from "./Dialog.vue";
-
-export default TaskRolloutActionDialog;
