--- frontend/src/components/IssueV1/components/Dialog/index.ts
@@ -1,9 +1,0 @@
-import IssueReviewActionDialog from "./IssueReviewActionDialog";
-import IssueStatusActionDialog from "./IssueStatusActionDialog";
-import TaskRolloutActionDialog from "./TaskRolloutActionDialog";
-
-export {
-  IssueReviewActionDialog,
-  IssueStatusActionDialog,
-  TaskRolloutActionDialog,
-};
