+++ frontend/src/components/IssueV1/components/Panel/index.ts
@@ -1,0 +1,9 @@
+import IssueReviewActionPanel from "./IssueReviewActionPanel.vue";
+import IssueStatusActionPanel from "./IssueStatusActionPanel.vue";
+import TaskRolloutActionPanel from "./TaskRolloutActionPanel.vue";
+
+export {
+  IssueStatusActionPanel,
+  IssueReviewActionPanel,
+  TaskRolloutActionPanel,
+};
